a3x/skills/requirements.txt
markdown2
xhtml2pdf
requests
httpx
python-slugify # For creating safe filenames from titles 